Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,21 +1,21 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     char c;
     scanf("%d%d%c", &a,&b,&c);
     switch (c){
         case '+':
-        printf("%f",a+b);
+        printf("%d",a+b);
         break;
         case '-':
-        printf("%f",a-b);
+        printf("%d",a-b);
         break;
         case '*':
-        printf("%f",a*b);
+        printf("%d",a*b);
         break;
         case '/':
         if (b!=0)
-        printf("%f",a/b);
+        printf("%d",a/b);
         else
         printf("error");
         break;
